fix(chat): register templates when email service exists
