+++ modules/chat/src/routes/routes.ts
@@ -17,6 +17,7 @@ import { isArray, isNil } from 'lodash';
 import { status } from '@grpc/grpc-js';
 import { validateUsersInput, sendInvitations } from '../utils';
 import { InvitationRoutes } from './InvitationRoutes';
+import * as templates from '../templates';
 
 export class ChatRoutes {
 
@@ -26,6 +27,23 @@ export class ChatRoutes {
   constructor(readonly server: GrpcServer, private readonly grpcSdk: ConduitGrpcSdk) {
     this._routingManager = new RoutingManager(this.grpcSdk.router, server);
     this.invitationRoutes = new InvitationRoutes(this.grpcSdk, this._routingManager);
+  }
+
+  async registerTemplates() {
+    this.grpcSdk.config
+      .get('email')
+      .then((emailConfig: any) => {
+        const promises = Object.values(templates).map((template) => {
+          return this.grpcSdk.emailProvider!.registerTemplate(template);
+        });
+        return Promise.all(promises);
+      })
+      .then(() => {
+        console.log('Email templates registered');
+      })
+      .catch((e: Error) => {
+        console.error('Internal error while registering email templates');
+      });
   }
 
   async createRoom(call: ParsedRouterRequest): Promise<UnparsedRouterResponse> {
@@ -528,8 +546,6 @@ export class ChatRoutes {
     }
 
     if (config.explicit_room_joins.enabled) {
-      if (config.explicit_room_joins.send_email)
-        await this.invitationRoutes.registerTemplates();
       this.invitationRoutes.declareRoutes();
     }
